Split SvgOffset
