+++ Source/Filter Effects/feOffset/SvgOffset.cs
@@ -1,5 +1,3 @@
-using System.Drawing;
-
 namespace Svg.FilterEffects
 {
     /// <summary>
@@ -34,27 +32,6 @@ namespace Svg.FilterEffects
             get { return _dy; }
             set { _dy = value; Attributes["dy"] = value; }
         }
-
-#if !NO_SDC
-        public override void Process(ImageBuffer buffer)
-        {
-            var inputImage = buffer[this.Input];
-            var result = new Bitmap(inputImage.Width, inputImage.Height);
-
-            var pts = new PointF[] { new PointF(this.Dx.ToDeviceValue(null, UnitRenderingType.Horizontal, null),
-                this.Dy.ToDeviceValue(null, UnitRenderingType.Vertical, null)) };
-            buffer.Transform.TransformVectors(pts);
-
-            using (var g = Graphics.FromImage(result))
-            {
-                g.DrawImage(inputImage, new Rectangle((int)pts[0].X, (int)pts[0].Y,
-                        inputImage.Width, inputImage.Height),
-                    0, 0, inputImage.Width, inputImage.Height, GraphicsUnit.Pixel);
-                g.Flush();
-            }
-            buffer[this.Result] = result;
-        }
-#endif
 
         public override SvgElement DeepCopy()
         {
